fix warnings
Fix a number of warnings so we can use meson warning level 3.  Instances
of the following warnings are found and resolved:

  Wreorder
  Wunused-parameter
  Wcatch-value
  Wsign-compare
  Waddress
  Wtautological-compare
  Wunused-but-set-variable

+++ include/CPUSensor.hpp
@@ -23,7 +23,7 @@ class CPUSensor : public Sensor
     boost::asio::posix::stream_descriptor inputDev;
     boost::asio::deadline_timer waitTimer;
     boost::asio::streambuf readBuf;
-    int errCount;
+    size_t errCount;
     void setupRead(void);
     void handleResponse(const boost::system::error_code &err);
     void checkThresholds(void) override;
