+++ include/TachSensor.hpp
@@ -64,7 +64,7 @@ class TachSensor : public Sensor
     boost::asio::posix::stream_descriptor inputDev;
     boost::asio::deadline_timer waitTimer;
     boost::asio::streambuf readBuf;
-    int errCount;
+    size_t errCount;
     void setupRead(void);
     void handleResponse(const boost::system::error_code &err);
     void checkThresholds(void) override;
